# Conflicts:
#	src/main/java/com/mp/karental/configuration/SecurityConfig.java
#	src/main/java/com/mp/karental/constant/ETransactionType.java
#	src/main/java/com/mp/karental/exception/ErrorCode.java

+++ src/main/java/com/mp/karental/entity/Car.java
@@ -54,10 +54,10 @@ public class Car {
     float fuelConsumption;
 
     @Column(name = "base_price", nullable = false)
-    int basePrice;
+    long basePrice;
 
     @Column(nullable = false)
-    int deposit;
+    long deposit;
 
     //address
     @Column(name = "city_province",nullable = false)
